requirements.txt
Django==4.2.5
djangorestframework==3.14.0
drf-yasg
mysqlclient
pillow
django-filter
stripe
python-dotenv
celery
redis
django-otp
pyotp
